requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
gcloud==0.18.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
numpy==1.25.2
oauth2client==4.1.3
pandas==2.0.3
protobuf==4.24.0rc2
pyarrow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
